+++ frontend/src/router/index.js
@@ -1,5 +1,4 @@
 import { createRouter, createWebHistory } from 'vue-router'
-
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
@@ -28,5 +27,4 @@ const router = createRouter({
     },
   ],
 })
-
 export default router